+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## 1.0.0
+
+### Added
+
+Added constructs:
+
+* **CloudFormation**
+  + `CloudFormationCustomResource` (*AWS:: CloudFormation:: CustomResource*)
+* **CloudFront**
+  + `CloudFrontDistribution` (*AWS:: CloudFront:: Distribution*)
+  + `CloudFrontFunction` (*AWS:: CloudFront:: Function*)
+* **CodeBuild**
+  + `CodeBuildSourceCredentials` (*AWS:: CodeBuild:: SourceCredential*)
+  + `CodeBuildProject` (*AWS:: CodeBuild:: Project*)
+* **Custom**
+  + `CustomResource` (*Custom:: AWS*)
+* **IAM**
+  + `IAMRole` (*AWS:: IAM:: Role*)
+  + `IAMPolicy` (*AWS:: IAM:: Policy*)
+* **Lambda**
+  + `LambdaFunction` (*AWS:: Lambda:: Function*)
+* **Route53**
+  + `Route53RecordSet` (*AWS:: Route53:: RecordSet*)
+* **S3**
+  + `S3Bucket` (*AWS:: S3:: Bucket*)
+  + `S3BucketPolicy` (*AWS:: S3:: BucketPolicy*)
+* **SSM**
+  + `SSMParameter` (*AWS:: SSM:: Parameter*)
+* **WafV2**
+  + `WafV2WebACL` (*AWS:: WAFv2:: WebACL*)
+
 ## 0.0.1 
 
 ### Added